+++ src/main/java/com/tobiasbrandy/challenge/meli1/resources/projections/PlainSatelliteProj.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import com.tobiasbrandy.challenge.meli1.models.Satellite;
 
-public record PlainSatelliteProj(String name, long positionX, long positionY) {
+public record PlainSatelliteProj(String name, double positionX, double positionY) {
     public static PlainSatelliteProj fromSatellite(final Satellite sat) {
         return new PlainSatelliteProj(sat.name(), sat.positionX(), sat.positionY());
     }
